Add selectivity estimate for CTID system variables
Previously, estimates for ItemPointer range quals, such as "ctid <= '(5,7)'",
resorted to the default value of 0.33 for range selectivity, although there was
special-case handling for equality quals like "ctid = '(5,7)'", which used the
appropriate selectivity for distinct items.

The estimator will now use the relation size to estimate the selectivity of a range qual.